refactor: move all ts2 utils to ts2 language
BREAKING CHANGE: all ts2 utils were moved to ts2 language
