+++ src/schema/2.0/operation-object.ts
@@ -1,7 +1,7 @@
 import { Option } from 'fp-ts/lib/Option';
 import { ExternalDocumentationObject } from './external-documentation-object';
 import { ParameterObject } from './parameter-object/parameter-object';
-import { Reference, ReferenceObject } from './reference-object';
+import { ReferenceObject } from './reference-object';
 import { ResponsesObject } from './responses-object';
 import { SecurityRequirementObject } from './security-requirement-object';
 import { booleanOption, stringArrayOption, stringOption } from '../../utils/io-ts';
@@ -16,7 +16,7 @@ export interface OperationObject {
 	readonly operationId: Option<string>;
 	readonly consumes: Option<string[]>;
 	readonly produces: Option<string[]>;
-	readonly parameters: Option<Reference<ParameterObject>[]>;
+	readonly parameters: Option<Array<ReferenceObject | ParameterObject>>;
 	readonly responses: ResponsesObject;
 	readonly schemes: Option<string[]>;
 	readonly deprecated: Option<boolean>;
