+++ src/schema/2.0/paths-object.ts
@@ -1,7 +1,7 @@
 import { Dictionary } from '../../utils/types';
-import { PathItemObject } from './path-item-object';
+import { PathItemObject, PathItemObjectCodec } from './path-item-object';
 import { dictionary } from '../../utils/io-ts';
 
 export interface PathsObject extends Dictionary<PathItemObject> {}
 
-export const PathsObject = dictionary(PathItemObject, 'PathsObject');
+export const PathsObject = dictionary(PathItemObjectCodec, 'PathsObject');
