+++ src/schema/2.0/reference-object.ts
@@ -4,8 +4,6 @@ export interface ReferenceObject {
 	readonly $ref: string;
 }
 
-export type Reference<A> = A | ReferenceObject;
-
 export const ReferenceObject = type(
 	{
 		$ref: string,
